backend/requirements.txt
fastapi
python-arango==7.5.4
uvicorn
pandas
paramiko
